separated validation to the new class Validator
